SGF-392 - Add support for OQL Query statement extensions in Repository Query methods via Annotations.
Adding integration tests to test the use of the custom OQL Query Annotation extensions in Repository Query methods against GemFire's actual OQL Query engine.
(cherry picked from commit bdbf101fd7f3e920f28634bb24cd0667855fa1e2)

+++ src/test/java/org/springframework/data/gemfire/repository/sample/UserRepositoryQueriesIntegrationTest.java
@@ -85,7 +85,7 @@ public class UserRepositoryQueriesIntegrationTest {
 	}
 
   protected static int toIntValue(final Integer value) {
-    return (value == null ? 0 : value.intValue());
+    return (value == null ? 0 : value);
   }
 
 	@Before
